ダイアログのコンポーネント化対応
